문제아
2. 위젯 컴포넌트랑 다시 연결
3. 데미지는 옳게 들어감 CurrentHP가 금쪽이임

+++ N_Graduation_project/Source/N_Graduation_project/Public/EntityWidget.h
@@ -12,7 +12,7 @@ class N_GRADUATION_PROJECT_API UEntityWidget : public UUserWidget
 public:
 	// HP 업데이트
 	UFUNCTION(BlueprintCallable, Category = "UI")
-	void UpdateHealthBar(int32 NewHealth);
+	void UpdateHealthBar(float NewHealth);
 
 	// 이름 받기
 	UFUNCTION(BlueprintCallable, Category = "Entity")
